Add support for the Unity-based port (WIP, untested)
